code review changes and made own component
