+++ src/RestaurantDisplay/index.js
@@ -2,7 +2,7 @@ import './styles.scss';
 const RestaurantDisplay = ({ restaurant }) => {
     return (
         <>
-            <div class="restaurant col-lg-3 col-12">
+            <div className='restaurant col-lg-3 col-12'>
                 <h1>{ restaurant.name }</h1>
             </div>
         </>
